Introduce `python_example` application (#6)
Initial commit

This introduces an example Python application using Django.

The application is similar in style and context to the original Rails
application, but a la Django instead.

There isn't quite feature parity, but I think that is OK.

Each of these applications is intended to give candidates a foundation from
which to pair on, and I think this does the job.

Co-authored-by: Sidra <sidra@ophelos.com>